% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_doc.R
\docType{data}
\name{anzsic_4_abs}
\alias{anzsic_4_abs}
\title{Australian New Zealand Standard Industry Classification}
\format{A data frame with 502 rows and 12 variables.}
\source{
ABS, courtesy of the Nous Group data warehouse
}
\description{
Price index of whole milk powder from July 2008 to August 2016.  Sometimes monthly, sometimes fortnightly.
}
