% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcp.R
\name{bcp}
\alias{bcp}
\title{bulk upload of a pipe delimited text file to an existing table in a SQL Survey database}
\usage{
bcp(
  server,
  database,
  schema,
  table,
  file,
  user = NULL,
  pwd = NULL,
  verbose = FALSE,
  extra_args = " -b 1000000",
  delim = "|"
)
}
\arguments{
\item{server}{address of a SQL Server server eg 'localhost' or a URL}

\item{database}{name of the database on the server to which to upload the data}

\item{schema}{name of the schema on the database to which to upload the data}

\item{table}{name of the (already existing) table on the database to upload the data}

\item{file}{name of a pipe-delimited text file to upload to the table, with no column headers}

\item{user}{user name if needed to write to the server}

\item{pwd}{password if needed to write to the server}

\item{verbose}{whether to print out the commands sent to the command line}

\item{extra_args}{extra arguments to pass through to the \code{bcp} utility}

\item{delim}{delimiter in the text file we are uploading}
}
\description{
bulk upload of a pipe delimited text file to an existing table in a SQL Survey database
}
\details{
This is a convenience wrapper for the \code{bcp} function for bulk upload to SQL Server.

Never type your password into a script; use \code{rstudioapi::askForPassword("Database password")} 
instead and pass it to the \code{pwd} argument.

I recommend you use something like
 \code{data.table:fwrite(your_data, "staging/your_data.txt", sep = "|", col.names = FALSE)}
as the fastest way to write your data to a pipe-delimited text file with no column names.

If the columns in the file you have saved don't match those in the table, this operation will fail,
possibly with an error, possibly silently. So take care.
}
\author{
Peter Ellis
}
