% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_doc.R
\name{build_doc}
\alias{build_doc}
\title{Build an R Markdown file that is in a subdirectory}
\usage{
build_doc(subdir = "report", report_name = "report", ...)
}
\arguments{
\item{subdir}{sub directory (relative to the project directory) holding a report and its necessary files}

\item{report_name}{name of the report, excluding the .Rmd suffix.}

\item{...}{other arguments to be passed to rmarkdown::render()}
}
\description{
Function to move all files in a subdirectory to the user's home directory, build the markdown file there,
and return the built version to the original subdirectory.  Why would you want to do this?  Because
R Markdown doesn't build nicely on RStudio Server on a mapped network drive due to a bug in Pandoc, so
sometimes you need to move everything to somewhere else.
}
\author{
Peter Ellis
}
