% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup.R
\name{cleanup}
\alias{cleanup}
\title{Function for removing all objects except those designated as 'keepers'}
\usage{
cleanup(keepers = NULL, verbose = FALSE)
}
\arguments{
\item{keepers}{a vector of the names of objects to keep}

\item{verbose}{verbosity of messages}
}
\description{
Function for removing all objects except those designated as 'keepers'
}
\author{
Peter Ellis
}
