% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps_doc.R
\docType{data}
\name{country_borders_tb}
\alias{country_borders_tb}
\title{Pacific Island exclusive economic zones}
\format{
A tibble with country borders of the world, suitable for using with Pacific-centred maps
}
\source{
Pacific Data Hub
}
\description{
Exclusive economic zones of the Pacific Island countries and territories that are members of the Pacific community
}
