% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_doc.R
\docType{data}
\name{dairy}
\alias{dairy}
\title{Whole Milk Powder Price Indices}
\format{A data frame with 170 rows and 2 variables.}
\source{
Global Dairy Trade database via Qandl
}
\description{
Price index of whole milk powder from July 2008 to August 2016.  Sometimes monthly, sometimes fortnightly.
}
