% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_sql.R
\name{datetime_ch}
\alias{datetime_ch}
\title{Helper function to convert the output of Sys.time() into a character string
without the time zone on it}
\usage{
datetime_ch(dt)
}
\arguments{
\item{dt}{an object of class \code{POSIXCT}}
}
\description{
Helper function to convert the output of Sys.time() into a character string
without the time zone on it
}
\details{
Not exported.
}
\examples{
datetime_ch(Sys.time())
}
\keyword{internal}
