% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulus-transform.R
\name{.mod_inverse}
\alias{.mod_inverse}
\title{Modulus transform}
\usage{
.mod_inverse(yt, lambda)
}
\arguments{
\item{yt}{variable to be inverse-transformed}

\item{lambda}{parameter passed to the Box-Cox transformation}
}
\description{
Convenience version of the inverse of this modulus transform function
}
\author{
Peter Ellis
}
