% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulus-transform.R
\name{.mod_transform}
\alias{.mod_transform}
\title{Modulus transform}
\usage{
.mod_transform(y, lambda)
}
\arguments{
\item{y}{variable to be transformed}

\item{lambda}{parameter passed to the Box-Cox transformation}
}
\description{
Convenience version of this modulus transform function
}
\author{
Peter Ellis
}
