% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_pac_map.R
\name{draw_pac_map}
\alias{draw_pac_map}
\title{Draw map of Pacific Island countries and territories}
\usage{
draw_pac_map(
  fill_df = NULL,
  join_col = "geo_pict",
  fill_col = NULL,
  xlim = c(120, 240),
  ylim = c(-50, 50),
  idl_col = "steelblue",
  leg_pos = c(0.8, 0.7)
)
}
\arguments{
\item{fill_df}{(Optional) A data frame with data to be used for colouring in
the EEZs}

\item{join_col}{Name of a column in fill_df to join to pac_map_sf. Should be
either geo_pict, iso3, or name2, and needs to match exactly the contents of
those columns in pac_map_sf}

\item{fill_col}{Column in fill_df to map to colour}

\item{xlim}{longitude to limit the map to}

\item{ylim}{latitude to limt the map to}

\item{idl_col}{Colour to use for the International date line}

\item{leg_pos}{Legend position (in grid units i.e. 0,0 is bottom left, 1,1 is top right)}
}
\value{
A ggplot2 object
}
\description{
Draw map of Pacific Island countries and territories
}
\details{
draws a map of Padcific Island countries and territories EEZs, with
  possibility of drawing it as a choropleth map.
}
