% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualplot.R
\name{dualplot}
\alias{dualplot}
\title{Dual scale time series plot}
\usage{
dualplot(
  x1,
  y1,
  y2,
  x2 = x1,
  col = c("#C54E6D", "#009380"),
  lwd = c(1, 1),
  colgrid = NULL,
  mar = c(3, 6, 3, 6) + 0.1,
  ylab1 = paste(substitute(y1), collapse = ""),
  ylab2 = paste(substitute(y2), collapse = ""),
  nxbreaks = 5,
  yleg1 = paste(gsub("\\n$", "", ylab1), "(left axis)"),
  yleg2 = paste(ylab2, "(right axis)"),
  ylim1 = NULL,
  ylim2 = NULL,
  ylim.ref = NULL,
  xlab = "",
  main = NULL,
  legx = "topleft",
  legy = NULL,
  bty = "n",
  silent = FALSE,
  ...
)
}
\arguments{
\item{x1}{x coordinates for first series}

\item{y1}{y coordinates for first series}

\item{y2}{y coordinates for second series}

\item{x2}{x coordinates for second series.  Defaults to be the same as x1 (ie both series covering the same time periods)}

\item{col}{vector of two colours for the two series}

\item{lwd}{vector of two line widths (as in \code{par})}

\item{colgrid}{colour of gridlines to be drawn (if NULL, none are drawn)}

\item{mar}{argument passed through to \code{par}}

\item{ylab1}{label for left y axis, for first series}

\item{ylab2}{label for right y axis, for second series}

\item{nxbreaks}{number of breaks to split the x axis into}

\item{yleg1}{label for first series in the legend}

\item{yleg2}{label for second series in the legend}

\item{ylim1}{Override \code{ylim} passed to \code{plot} for the first series.  Not recommended.}

\item{ylim2}{Override \code{ylim} passed to \code{plot} for the second series.  Not recommended.}

\item{ylim.ref}{vector of two, indicating which element in the first series and the second series to use as the reference for converting to an index for purpose of determining ylim.  Only used if all y1 and y2 are positive.}

\item{xlab}{label for x (horizontal) axis}

\item{main}{main title for plot}

\item{legx}{passed through to \code{legend}, controls position of legend}

\item{legy}{passed through to \code{legend}}

\item{bty}{passed through to \code{legend} 'o' or 'n' indicates whether or not (respectively) to draw a box around the legend}

\item{silent}{suppress warnings and messages?}

\item{...}{other arguments passed through to \code{legend}}
}
\description{
Time series line chart with two different vertical scales
}
\details{
Draws a time series line chart with two different vertical scales.  If defaults are used, \code{dualplot()} 
attempts to derive sensible values of ylim1 and ylim2 that minimise the misinterpretation common to dual axis time
series charts.  In particular:
\enumerate{
 \item if all values of y1 and y2 are > 0, the y limits are chosen so the two series are drawn as though they had been converted to indexes, using the \code{ylim.ref[1]}th element of y1  and \code{ylim.ref[2]}th element of y2 as their reference points.
   \itemize{
     \item If y1 and y2 are equal lengths, ylim.ref defaults to 1, 1
     \item If y1 and y2 are unqual lengths, ylim.ref takes the values that make the two series cross over at the point in time where the later starting series begins.
 }
 \item if either y1 or y2 goes to zero or below, the y limits are chosen to be the mean of each series +/- 3.5 standard deviations
}

The data should be sorted (earliest to most recent) before use or odd choices might be made.
}
\examples{

data(dairy)
data(fonterra)
# with defaults for positioning scales, but exemplifying custom colours, etc:
dualplot(x1 = dairy$Date, y1 = dairy$WMP_Total, 
         x2 = fonterra$Date, y2 = fonterra$FCGClose,
         ylab1 = "Whole milk powder price index\n",
         ylab2 = "Fonterra Cooperative Group\nshare price ($)",
         col = c("#AA6800", "#7869CD"),
         colgrid = "grey90",
         main = "Fonterra share prices are less volatile than global milk prices")

# draw as though each series indexed, starting from its first point:
dualplot(x1 = dairy$Date, y1 = dairy$WMP_Total, 
         x2 = fonterra$Date, y2 = fonterra$FCGClose,
         ylim.ref = c(1, 1))

# draw as though each series indexed, ending at the same point:
dualplot(x1 = dairy$Date, y1 = dairy$WMP_Total, 
         x2 = fonterra$Date, y2 = fonterra$FCGClose,
         ylim.ref = c(nrow(dairy), nrow(fonterra)))

# pretend one of the series is negative and see how the axes are chosen differently:
dualplot(x1 = dairy$Date, y1 = -dairy$WMP_Total, 
         x2 = fonterra$Date, y2 = fonterra$FCGClose)
# note the plot above is *not* meaningful - just for illustrating negatives.
}
\references{
\url{http://ellisp.github.io/blog/2016/08/18/dualaxes}, \url{http://ellisp.github.io/blog/2016/08/28/dualaxes2}
}
\seealso{
\code{\link{par}}
}
\author{
Peter Ellis, \email{peter.ellis2013nz@gmail.com}
}
