% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo_rating.R
\name{elo_rating}
\alias{elo_rating}
\title{Determine Elo rating of two players given one has won a match}
\usage{
elo_rating(a, b, winner = c("a", "b"), ml = 1, axp = 500,
  bxp = 500)
}
\arguments{
\item{a}{Elo rating of player a before match}

\item{b}{Elo rating of player b before match}

\item{winner}{whether A or B is the winner}

\item{ml}{is match length}

\item{axp}{total match lengths (experience) of player a until this match}

\item{bxp}{total match lengths (experience) of player b until this match}
}
\description{
Determine Elo rating of two players given one has won a match
}
\details{
See http://www.fibs.com/ratings.html for formulae
}
\examples{
# baptism by fire; a's rating should be 1540.95 after beating a master player:
round(elo_rating(a = 1500, b = 1925, ml = 7, axp = 0, bxp = 10000, winner = "a")$a, 2)
}
\author{
Peter Ellis
}
