% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_doc.R
\docType{data}
\name{fonterra}
\alias{fonterra}
\title{Closing share price of Fonterra Cooperative Group}
\format{
A data frame with 976 rows and 2 variables.
}
\source{
New Zealand Stock Exchange via Yahoo Finance
}
\description{
Daily closing share price in New Zealand dollars of Fonterra Cooperative Group shares from 30 November 2012
to 29 August 2016.
}
