% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/death_rates.R
\docType{data}
\name{french_death_rates_2015}
\alias{french_death_rates_2015}
\title{French death rates 2015}
\format{
An object of class \code{data.frame} with 19 rows and 3 columns.
}
\source{
https://www.ined.fr/en/everything_about_population/data/france/deaths-causes-mortality/mortality-rates-sex-age/ on 27 May 2018
}
\usage{
french_death_rates_2015
}
\description{
Death rates per thousand for France in 2015
}
\examples{
life_expectancy(data = french_death_rates_2015[ , c("age", "female")])
life_expectancy(age = french_death_rates_2015$age, rate = french_death_rates_2015$male)
}
\keyword{datasets}
