% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frs.R
\docType{package}
\name{frs-package}
\alias{frs}
\alias{frs-package}
\title{Free Range Statistics}
\description{
Miscellaneous odds and ends associated with the Free Range Statistics blog
}
\details{
These are items that have been used in writing posts for \link{http://freerangestats.info} or at least
thinking about writing posts for it.
 
So far, the bits and pieces include:
\itemize{
    \item{dual axis plots}
    \item{so-called "modulus" transformation for ggplot2 scales, which is like a Box Cox of the 
    absolute value then with the sign returned.  Excellent for visualising economic variables that 
    can be zero or negative}
    \item{French death rates data in 2015}
    \item{Colour palette and ggplot2 theme for International Finance Corporation branding 
    (disclaimer - not associated at all with the IFC, but I did use it successfully in a 
    report for them once)}
    \item{turn levels of a factor into letters, useful for quick and dirty confidentialisation}
    \item{some ODBC database odds and ends to help with R interacting with SQL Server}
}
}
\author{
\strong{Maintainer}: Peter Ellis \email{peter.ellis2013nz@gmail.com}

}
