% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pdh_codelists.R
\name{get_pdh_codelists}
\alias{get_pdh_codelists}
\title{Get code lists from Pacific Data Hub PDH.Stat}
\usage{
get_pdh_codelists(dataflow = NULL, version = "1.0", str_url = NULL)
}
\arguments{
\item{dataflow}{character string name of a data flow eg DF_EMPRATES}

\item{version}{version number}
}
\description{
Get code lists from Pacific Data Hub PDH.Stat
}
\examples{
eg1 <- get_pdh_codelists(dataflow = "DF_EMPRATES", version = "1.0")
eg1
}
