% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifc-branding.R
\docType{data}
\name{ifc_pal}
\alias{ifc_pal}
\title{IFC palette}
\format{An object of class \code{data.frame} with 9 rows and 7 columns.}
\usage{
ifc_pal
}
\description{
Official palette for use for the International Finance Corporation of the World Bank Group
}
\examples{
plot(1:9, 1:9, pch = 19, col = ifc_pal$hex, cex = 8)
}
\keyword{datasets}
