% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps_doc.R
\docType{data}
\name{international_date_line_sf}
\alias{international_date_line_sf}
\title{International date line}
\format{
An sf lines object
}
\source{
Natural Earth
}
\description{
International date line
}
