% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/letterize.R
\name{letterize}
\alias{letterize}
\title{Letterize}
\usage{
letterize(x, method = c("random", "original", "frequency"),
  output = c("original", "character", "factor"))
}
\description{
Convert levels of a factor or character to letters, for anonymisation
}
\examples{
x <- c("Peter", "John", "Mary", "Jane", "John")
letterize(x)
letterize(x, output = "factor")
letterize(x, method = "frequency")
letterize(x, method = "original") # alphabetical
}
