% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/death_rates.R
\name{life_expectancy}
\alias{life_expectancy}
\title{Life expectancy}
\usage{
life_expectancy(age = data[, 1], rate = data[, 2], data = NULL)
}
\arguments{
\item{age}{vector of ages in years}

\item{rate}{death rate in deaths per 1000 people alive at the given age.  Must be the same length as age.}

\item{data}{optionally, a data frame or matrix where the first column is used as age and the second column as rate}
}
\description{
Calculate period life expectancy at birth, given a set of death rates at various ages
}
\details{
Period LEB is the mean length of life of a hypothetical cohort assumed to be exposed, from birth through death, 
to the mortality rates observed at a given year.  So for the person born in year X, their death rate at age 10 is
assumed to be the death rate as it is in year X; there is no attempt to forecast the death rate for ten year olds
in year X + 10.  Hence if death rates are in decline, period life expectancy at birth will be an underestimate of
actual cohort life expectancy which can be measured when all members of the cohort are dead (and vice versa if death
rates are increasing).
}
\examples{
life_expectancy(data = french_death_rates_2015[ , c("age", "female")])
life_expectancy(age = french_death_rates_2015$age, rate = french_death_rates_2015$male)
}
\references{
https://en.wikipedia.org/wiki/Life_expectancy
}
