% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulus-transform.R
\name{modulus_breaks}
\alias{modulus_breaks}
\title{Modulus transform breaks}
\usage{
modulus_breaks(lambda, n = 8, prettify = TRUE)
}
\arguments{
\item{lambda}{parameter for modulus transform}

\item{n}{number of axis labels expected}

\item{prettify}{whether or not to prettify the axis labels}
}
\value{
a function to calculate breaks, suitable for use in conjunction with the scales package
}
\description{
Create a function to calculate sensible breaks for axis labels in combination with a modulus transform
}
\examples{
eg_data <- data.frame(x = exp(rnorm(1000)) * 
               sample(c(-1, 1), 1000, replace = TRUE, prob = c(0.2, 0.8)))

ggplot(eg_data, aes(x = x)) +
  geom_density() +
  scale_x_continuous("Transformed scale",
                     trans = modulus_trans(0.1),
                     breaks = modulus_breaks(lambda = 0.1)) 
}
\references{
Inspired by Heather Turner's answer to a Stack Overflow question at 
\url{https://stackoverflow.com/questions/14255533/pretty-ticks-for-log-normal-scale-using-ggplot2-dynamic-not-manual}
}
\seealso{
\code{\link{prettify}}, \code{\link{modulus_trans}}
}
\author{
Peter Ellis
}
