% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulus-transform.R
\name{modulus_trans}
\alias{modulus_trans}
\title{Modulus transform}
\usage{
modulus_trans(lambda)
}
\arguments{
\item{lambda}{parameter to pass to the underlying modulus transformation}
}
\value{
An object of class "trans", which encapsulates the transformation and its inverse
}
\description{
John and Draper's modulus transformation
}
\details{
Creates a function to execute a Box-Cox power transformation to the absolute value of a variable and then restores its sign.
}
\examples{
eg_data <- data.frame(x = exp(rnorm(1000)) * 
               sample(c(-1, 1), 1000, replace = TRUE, prob = c(0.2, 0.8)))

ggplot(eg_data, aes(x = x)) +
  geom_density() +
  scale_x_continuous("Transformed scale",
                     trans = modulus_trans(0.1),
                     breaks = modulus_breaks(lambda = 0.1))
}
\references{
J. A. John and N. R. Draper, "An Alternative Family of Transformation", 
\emph{Journal of the Royal Statistical Society. Series C (Applied Statistics)} 
Vol. 29, No. 2 (1980) pp. 190-197

\url{http://ellisp.github.io/blog/2015/09/05/creating-a-scale-transformation}
}
\author{
Peter Ellis
}
