% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps_doc.R
\docType{data}
\name{pac_map_sf}
\alias{pac_map_sf}
\title{Pacific Island exclusive economic zones}
\format{
An sf data frame
}
\source{
Pacific Data Hub
}
\description{
Exclusive economic zones of the Pacific Island countries and territories that are members of the Pacific community
}
