% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulus-transform.R
\name{prettify}
\alias{prettify}
\title{Prettify}
\usage{
prettify(x)
}
\arguments{
\item{x}{numbers to be rounded}
}
\value{
Rounded numbers
}
\description{
Turn a set of numbers into a rounded set
}
\details{
Rounding is more aggressive the larger the numbers are.  Expected usage is for axis labels or similar.
}
\examples{
prettify(c(1, 8.3, 98.78987))
}
