% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_all_r_scripts.R
\name{run_all_r_scripts}
\alias{run_all_r_scripts}
\title{Run all R scripts in a folder system}
\usage{
run_all_r_scripts(
  path,
  recursive = TRUE,
  verbose = TRUE,
  push_through = FALSE,
  cleanup = TRUE,
  suppress_messages = TRUE
)
}
\arguments{
\item{path}{relative file path to the folder with R scripts to run}

\item{recursive}{whether or not to run all scripts in sub-folders as well as in the path}

\item{verbose}{whether or not to print to the screen the name of each script as run}

\item{push_through}{whether to push through to the next script if an error is found (otherwise will stop)}

\item{cleanup}{whether to remove all objects from the global environment other than those designated as keepers.
If there is an object called \code{keepers} in the global environment, then any objects named in \code{keepers}
will not be removed at the end of running each script.}

\item{suppress_messages}{whether messages should be suppressed or not. Sometimes messages can be overwhelming.}
}
\description{
Run all R scripts in a folder system
}
\author{
Peter Ellis
}
