% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifc-branding.R
\name{scale_color_discrete_ifc}
\alias{scale_color_discrete_ifc}
\title{Discrete color scale for the IFC}
\usage{
scale_color_discrete_ifc(..., sequence = c(4, 5, 9, 6, 7, 8))
}
\arguments{
\item{...}{parameters to be passed to \code{scale_color_manual()}}

\item{sequence}{sequence of colors from \code{ifc_pal} to use}
}
\description{
Discrete color scale for the IFC
}
\seealso{
\code{\link{ifc_pal}}
}
