% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifc-branding.R
\name{scale_fill_ifc}
\alias{scale_fill_ifc}
\title{Fill scale for the IFC}
\usage{
scale_fill_ifc(..., sequence = c(4, 5, 9, 6, 7, 8))
}
\arguments{
\item{...}{parameters to be passed to \code{scale_fill_manual()}}

\item{sequence}{sequence of colors from \code{ifc_pal} to use}
}
\description{
Fill scale for the IFC
}
\seealso{
\code{\link{ifc_pal}}
}
