% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution.R
\name{sharpen}
\alias{sharpen}
\title{Single dimensional deconvolution}
\usage{
sharpen(y, cvv, warnings = FALSE, digits = NULL)
}
\arguments{
\item{y}{a vector of values to be deconvolved}

\item{cvv}{a vector of probabilities for the original convolution that created y}

\item{warnings}{passed through to blur}

\item{digits}{how many digits to round the deconvolved values to. If NULL no rounding occurs.}
}
\value{
a data frame with columns for x (the inferred original values what were convolved to y),
y (which will be padded out with some extra zeroes), and position (which is the numbering of the
row relative to the original ordering of y; so position = 1 refers to the first value of y)
}
\description{
Single dimensional deconvolution
}
\details{
This is simply the inverse of blur. However there is no closed solution so the
result is estimated numerically. The column x from the data frame returned by this function is
a set of values that, if blur is applied to it with the cvv vector of probabilities, will
return y.
}
\examples{
x <- c(4,6,8,9,7,5)
cvv <- c(0, 0.3, 0.5, 0.2)
# create a convolved version of x:
y <- blur(x, cvv)
# recover the original version of x, given its blurred version 
# and the original convolution probabilities:
sharpen(y, cvv)
}
