% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg_googlefonts.R
\name{svg_googlefonts}
\alias{svg_googlefonts}
\title{Add Google webfonts to the style definition of an SVG file}
\usage{
svg_googlefonts(
  svgfile,
  googlefonts = c("Roboto", "Sarala"),
  new_svgfile = svgfile
)
}
\arguments{
\item{svgfile}{path to an SVG}

\item{googlefonts}{Google Web fonts to be added}

\item{new_svgfile}{path to a new SVG file to create. By default will overwrite \code{svgfile}.}
}
\description{
Add Google webfonts to the style definition of an SVG file
}
\details{
This is a very quick hack put together for my purposes, I don't know how generally well 
it will work.
}
\examples{
\dontrun{
library(svglite)
svglite("test.svg")
par(family =  "Indie Flower")
plot(1:10, 1:10, main = "Indie Flower", xlab = "X axis here")
text(10:1, 1:10, LETTERS[1:10])
dev.off()
svg_googlefonts("test.svg", "Indie Flower")
}
}
\author{
Peter Ellis
}
