% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg_png.R
\name{svg_png}
\alias{svg_png}
\title{Draw a graphic as both an SVG and a PNG}
\usage{
svg_png(p, file, w = 8, h = 5, res = 600, googlefonts = c("Roboto",
  "Sarala"))
}
\arguments{
\item{p}{ggplot2 object to draw, or base plot as a function}

\item{file}{stub of filename under which to save the graphic. .svg and .png will be added automatically.}

\item{w}{width in inches}

\item{h}{height in inchest}

\item{res}{resolution in dots per inch, to be passed to \code{Cairo::CairoPNG}}

\item{googlefonts}{vector of names of Google Fonts to be linked to from the SVG, to be passed
to \code{frs::svg_googlefonts}}
}
\description{
Draw a graphic as both an SVG and a PNG
}
\details{
Under the hood, this uses CairoPNG, svglite, and svg_googlefonts to produce a high
quality SVG and an identical (hopefully) high quality PNG
}
\examples{
if(require(ggplot2)){
  p1 <- ggplot(mtcars, aes(x = disp, y = mpg)) + geom_point()
  svg_png(p1, "p1")
  }

p2 <- function(){plot(1:10, 1:10)}
svg_png(p2, "p2")

if(require(lattice)){
  Depth <- equal.count(quakes$depth, number=8, overlap=.1)
  p3 <- xyplot(lat ~ long | Depth, data = quakes)
  svg_png("p3", p3)
  }

}
