% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifc-branding.R
\name{theme_ifc}
\alias{theme_ifc}
\title{IFC theme}
\usage{
theme_ifc(base_size = 12, base_family = "Arial")
}
\arguments{
\item{base_size}{base size}

\item{base_family}{typeface}
}
\description{
A ggplot2 theme for the International Finance Corporation
}
