% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution.R
\name{blur}
\alias{blur}
\title{One-dimensional convolution}
\usage{
blur(x, pmf, warnings = TRUE, scale_pmf = FALSE)
}
\arguments{
\item{x}{a vector of counts or other numbers}

\item{pmf}{a vector of probabilities to delay x at various lags. First value should be for lag 0, 
second for lag 1, etc.}

\item{warnings}{whether to show warnings when the resulting vector does not add up to the
same sum as the original}

\item{scale_pmf}{whether or not to scale pmf so it adds exactly to one}
}
\value{
A vector of length equal to the length of x plus length of pmf minus 1
}
\description{
One-dimensional convolution
}
\details{
\code{blur} and \code{sharpen} are deterministic single dimensional convolution functions
for simple convolution by a lagged probability mass function representing the proportion of original
cases that are delayed at various lags. They are for illustrative / toy purposes and probably should
not be used for actual analysis.
}
\examples{
x <- c(4,6,8,9,7,5)
pmf <- c(0, 0.3, 0.5, 0.2)
blur(x, pmf)
}
