% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_pac_map.R
\name{draw_pac_map}
\alias{draw_pac_map}
\title{Draw map of Pacific Island countries and territories}
\usage{
draw_pac_map(
  fill_df = NULL,
  join_col = "geo_pict",
  fill_col = NULL,
  fill_col_label = fill_col,
  base_size = 11,
  xlim = c(120, 240),
  ylim = c(-50, 50),
  country_labels = TRUE,
  country_label_size = 3,
  country_label_col = "black",
  idl_col = "steelblue",
  idl_label_size = country_label_size,
  leg_pos = c(0.8, 0.7),
  ocean_col = "lightsteelblue",
  family = "sans"
)
}
\arguments{
\item{fill_df}{(Optional) A data frame with data to be used for colouring in
the EEZs}

\item{join_col}{Name of a column in fill_df to join to pac_map_sf. Should be
either geo_pict, iso3, or name2, and needs to match exactly the contents of
those columns in pac_map_sf}

\item{fill_col}{Column in fill_df to map to colour}

\item{fill_col_label}{Label for the fill colour legend}

\item{base_size}{base size of fonts eg for title, passed on to theme_minimal()}

\item{xlim}{longitude to limit the map to}

\item{ylim}{latitude to limt the map to}

\item{country_labels}{whether or not to print the country names as labels (Palau, Guam, etc)}

\item{country_label_size}{Size of country labels. Set to zero to make them invisible.}

\item{country_label_col}{Colour of country labels.}

\item{idl_col}{Colour to use for the International date line}

\item{idl_label_size}{Size of the International date line label,}

\item{leg_pos}{Legend position (in grid units i.e. 0,0 is bottom left, 1,1 is top right)}

\item{ocean_col}{Colour to draw the ocean}

\item{family}{Font family}
}
\value{
A ggplot2 object
}
\description{
Draw map of Pacific Island countries and territories
}
\details{
draws a map of Padcific Island countries and territories EEZs, with
  possibility of drawing it as a choropleth map.
}
\examples{

draw_pac_map(country_label_size = 5)
}
