% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo_rating.R
\name{elo_prob}
\alias{elo_prob}
\title{Determine the expected probability of player a winning a match, given Elo ratings}
\usage{
elo_prob(a, b, ml)
}
\arguments{
\item{a}{Elo rating of player a}

\item{b}{Elo rating of player b}

\item{ml}{length of match (if a backgammon game, the minimum winning score)}
}
\description{
Determine the expected probability of player a winning a match, given Elo ratings
}
\examples{
# 50\% chance of winning if both players have equal ratings:
elo_prob(1500, 1500, 1)

# 87\% chance of winning a match to 17 if you're a backgammon expert v average player:
elo_prob(1900, 1500, 17)
}
