% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifc-branding.R
\name{scale_color_continuous_ifc}
\alias{scale_color_continuous_ifc}
\title{Continuous color scale for the IFC}
\usage{
scale_color_continuous_ifc(..., type = c("sequential", "diverging"),
  first_col = "red", second_col = "blue")
}
\description{
Continuous color scale for the IFC
}
\examples{
ggplot(mtcars, aes(x = vs, weight = mpg, fill = as.factor(cyl))) + 
 geom_bar() + 
 theme_ifc() +
 ggtitle("CHART TITLE", "SUBTITLE") +
 labs(caption = "Footer") +
 scale_fill_ifc()

ggplot(mtcars, aes(x = disp, y = mpg, color = mpg)) + 
 geom_point(size = 5) + 
 theme_ifc() +
 ggtitle("CHART TITLE", "SUBTITLE") +
 labs(caption = "Footer") +
 scale_color_continuous_ifc("diverging", second_col = 9)
}
