% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifc-branding.R
\name{scale_color_discrete_ifc}
\alias{scale_color_discrete_ifc}
\title{Discrete color scale for the IFC}
\usage{
scale_color_discrete_ifc(..., sequence = c(1, 3, 2, 4:9))
}
\description{
Discrete color scale for the IFC
}
