% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution.R
\name{sharpen}
\alias{sharpen}
\title{Single dimensional deconvolution}
\usage{
sharpen(y, pmf, warnings = FALSE, digits = NULL)
}
\arguments{
\item{y}{a vector of values to be deconvolved}

\item{pmf}{a vector of probabilities for the original convolution that created y}

\item{warnings}{passed through to blur}

\item{digits}{how many digits to round the deconvolved values to. If NULL no rounding occurs.}
}
\value{
a data frame with columns for x (the inferred original values what were convolved to y),
y (which will be padded out with some extra zeroes), and position (which is the numbering of the
row relative to the original ordering of y; so position = 1 refers to the first value of y)
}
\description{
Single dimensional deconvolution
}
\details{
\code{blur} and \code{sharpen} are deterministic single dimensional convolution functions
for simple convolution by a lagged probability mass function representing the proportion of original
cases that are delayed at various lags. They are for illustrative / toy purposes and probably should
not be used for actual analysis.

Use \code{\link[surveillance]{backprojNP}} for a better, maximum likelihood approach to recovering
an unseen set of original cases that result in the observations.

\code{sharpen} is the inverse of \code{blur}; it seeks to recover an original vector that, when blurred
via \code{pmf}, would produce the actual observations.
}
\examples{
x <- c(4,6,8,9,7,5)
pmf <- c(0, 0.3, 0.5, 0.2)
# create a convolved version of x:
y <- blur(x, pmf)
# recover the original version of x, given its blurred version 
# and the original convolution probabilities:
sharpen(y, pmf)
}
\seealso{
\code{\link[surveillance]{backprojNP}}.
}
